move to fileCrawler class
